@@ -1,86 +1,91 @@
 {
-  "name": "@moleculer/workflows",
-  "version": "0.0.0",
-  "description": "Advanced Workflow feature for Moleculer microservices framework",
-  "main": "index.js",
-  "scripts": {
-    "dev": "nodemon --inspect examples/index.js",
-    "dev:noservice": "nodemon --inspect=9228 examples/index.js simple noservice",
-    "ci:unit": "jest --testMatch \"**/unit/**/*.spec.js\" --watch",
-    "ci:integration": "jest --testMatch \"**/integration/**/*.spec.js\" --runInBand --watch",
-    "ci:leak": "jest --testMatch \"**/leak-detection/**/index.spec.js\" --runInBand --watch",
-    "test:unit": "jest --testMatch \"**/unit/**/*.spec.js\" --force-exit",
-    "test:integration": "jest --testMatch \"**/integration/**/*.spec.js\" --runInBand --coverage --force-exit",
-    "test:leak": "jest --testMatch \"**/leak-detection/**/*.spec.js\" --runInBand --force-exit",
-    "test": "jest --testMatch \"**/*.spec.js\" --runInBand --coverage --force-exit",
-    "lint": "eslint --ext=.js src examples test",
-    "bench": "node benchmark/index.js",
-    "bench:watch": "nodemon benchmark/index.js",
-    "deps": "ncu -i && npm audit fix",
-    "ci-update-deps": "ncu -u --target minor",
-    "check": "tsc --noEmit true",
-    "types": "tsc --emitDeclarationOnly true --noEmitOnError false",
-    "coverall": "cat ./coverage/lcov.info | ./node_modules/coveralls/bin/coveralls.js",
-    "release": "npm publish --access public && git push --tags"
-  },
-  "keywords": [
-    "workflow",
-    "moleculer",
-    "microservice"
-  ],
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/moleculerjs/workflows.git"
-  },
-  "author": "MoleculerJS",
-  "license": "MIT",
-  "typings": "index.d.ts",
-  "peerDependencies": {
-    "moleculer": "^0.14.12 || ^0.15.0-0"
-  },
-  "devDependencies": {
-    "axios": "^1.9.0",
-    "benchmarkify": "^4.0.0",
-    "eslint": "^8.57.0",
-    "eslint-config-prettier": "^9.1.0",
-    "eslint-plugin-node": "^11.1.0",
-    "eslint-plugin-prettier": "^5.2.3",
-    "eslint-plugin-promise": "^6.6.0",
-    "eslint-plugin-security": "^2.1.1",
-    "fakerator": "^0.3.6",
-    "globby": "^14.1.0",
-    "jest": "^29.7.0",
-    "jest-cli": "^29.7.0",
-    "kleur": "^4.1.5",
-    "moleculer": "^0.14.35",
-    "moleculer-repl": "^0.7.4",
-    "moleculer-web": "^0.10.8",
-    "msgpack5": "^6.0.2",
-    "nodemon": "^3.1.10",
-    "npm-check-updates": "^18.0.1",
-    "prettier": "^3.5.3",
-    "typescript": "^5.8.3"
-  },
-  "jest": {
-    "testEnvironment": "node",
-    "rootDir": "./src",
-    "roots": [
-      "../test"
-    ],
-    "coverageDirectory": "../coverage",
-    "coveragePathIgnorePatterns": [
-      "/node_modules/",
-      "/test/"
-    ]
-  },
-  "engines": {
-    "node": ">= 20.x.x"
-  },
-  "dependencies": {
-    "@types/lodash": "^4.17.17",
-    "cron-parser": "^5.1.1",
-    "fastest-validator": "^1.19.1",
-    "ioredis": "^5.6.1",
-    "lodash": "^4.17.21"
-  }
+	"name": "@moleculer/workflows",
+	"version": "0.0.0",
+	"description": "Advanced Workflow feature for Moleculer microservices framework",
+	"type": "module",
+	"main": "./dist/cjs/index.js",
+	"module": "./dist/esm/index.js",
+	"types": "./dist/types/index.d.ts",
+	"exports": {
+		".": {
+      "types": "./dist/types/index.d.ts",
+      "require": "./dist/cjs/index.js",
+      "import": "./dist/esm/index.js",
+      "default": "./src/index.ts"
+		}
+	},
+	"files": [
+		"README.md",
+		"LICENSE",
+		"dist"
+	],
+	"scripts": {
+		"dev": "nodemon --inspect examples/index.ts",
+		"dev:noservice": "nodemon --inspect=9228 examples/index.ts simple noservice",
+		"ci:unit": "vitest --watch unit",
+		"ci:integration": "vitest --watch integration",
+		"test:unit": "vitest --run unit",
+		"test:integration": "vitest --run integration  --coverage",
+		"test": "vitest --run --coverage",
+		"bench": "node benchmark/index.js",
+		"bench:watch": "nodemon benchmark/index.js",
+		"deps": "ncu -i && npm audit fix",
+		"ci-update-deps": "ncu -u",
+		"lint": "eslint --ext=.ts src examples test",
+		"check": "tsc --noEmit true",
+		"prebuild": "rimraf dist",
+		"build": "tsc -b tsconfig.cjs.json tsconfig.esm.json tsconfig.types.json",
+		"prepack": "npm run build",
+		"release": "npm publish --access public && git push --tags",
+		"postrelease": "git push origin && git push origin -f --tags"
+	},
+	"keywords": [
+		"workflow",
+		"moleculer",
+		"microservice"
+	],
+	"repository": {
+		"type": "git",
+		"url": "https://github.com/moleculerjs/workflows.git"
+	},
+	"author": "MoleculerJS",
+	"license": "MIT",
+	"peerDependencies": {
+		"moleculer": "^0.14.12 || ^0.15.0-0"
+	},
+	"dependencies": {
+		"cron-parser": "^5.2.0",
+		"ioredis": "^5.6.1",
+		"lodash": "^4.17.21"
+	},
+	"devDependencies": {
+		"@eslint/js": "^9.28.0",
+		"@types/lodash": "^4.17.17",
+		"@types/node": "^22.15.29",
+		"@vitest/coverage-v8": "^3.1.4",
+		"benchmarkify": "^4.0.0",
+		"eslint": "^9.28.0",
+		"eslint-config-prettier": "^10.1.5",
+		"eslint-plugin-node": "^11.1.0",
+		"eslint-plugin-prettier": "^5.4.1",
+		"eslint-plugin-promise": "^7.2.1",
+		"eslint-plugin-security": "^3.0.1",
+		"kleur": "^4.1.5",
+		"moleculer": "github:moleculerjs/moleculer#next",
+		"moleculer-repl": "^0.7.4",
+		"moleculer-web": "^0.10.8",
+		"msgpack5": "^6.0.2",
+		"nodemon": "^3.1.10",
+		"npm-check-updates": "^18.0.1",
+		"prettier": "^3.5.3",
+		"rimraf": "^6.0.1",
+		"tsup": "^8.5.0",
+		"tsx": "^4.19.4",
+		"typescript": "^5.8.3",
+		"typescript-eslint": "^8.33.0",
+		"vitest": "^3.1.4"
+	},
+	"engines": {
+		"node": ">= 20.x.x"
+	}
 }
